% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classifier.R
\name{classifier_network}
\alias{classifier_network}
\title{Custom classier network}
\usage{
classifier_network(n_classes, num_rois, pooling_regions)
}
\arguments{
\item{n_classes}{enter number of classes.}

\item{num_rois}{enter number of roi's.}

\item{pooling_regions}{enter number of pooling regions.}
}
\value{
creates a classifer network
}
\description{
Custom classier network
}
