% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{fasterrcnn}
\alias{fasterrcnn}
\title{Custom faster RCNN network}
\usage{
fasterrcnn(n_classes, num_rois, pooling_regions, num_anchors)
}
\arguments{
\item{n_classes}{enter number of classes. Ex 21}

\item{num_rois}{enter number of roi's. Ex 32}

\item{pooling_regions}{enter number of pooling regions. Ex 7}

\item{num_anchors}{enter number of anchors. Ex 4}
}
\value{
creates a classifer network
}
\description{
Custom faster RCNN network
}
\examples{
# fasterrcnn(n_classes=4, num_rois=32, pooling_regions=7, num_anchors=4)
}
