% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roi_pooling.R
\name{layer_roi_pooling_conv}
\alias{layer_roi_pooling_conv}
\title{RoI Pooling Layer}
\usage{
layer_roi_pooling_conv(object, pool_size, num_rois, name = NULL,
  trainable = TRUE)
}
\arguments{
\item{object}{Model or layer object}

\item{pool_size}{Size of pooling region to use. pool_size = 7 will result in
a 7x7 region.}

\item{num_rois}{number of regions of interest to be used}

\item{name}{An optional name string for the layer. Should be unique in a model
(do not reuse the same name twice). It will be autogenerated if it isn't
provided.}

\item{trainable}{Whether the layer weights will be updated during training.}
}
\description{
RoI Pooling Layer
}
