% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rpn.R
\name{region_proposal_network}
\alias{region_proposal_network}
\title{Region of Proposal Network}
\usage{
region_proposal_network(num_anchors)
}
\arguments{
\item{num_anchors}{enter number of anchors. Ex 4}
}
\value{
creates a classifer network
}
\description{
Region of Proposal Network
}
\examples{
# region_proposal_network(num_anchors=4)
}
